Fix #922 – GDAX improvements: round amounts to 8 decimals, have post_only default to true
Amounts that had more than 8 decimals resulted in an API error `Size too precise`. They're now trimmed.

`post_only: false` was suggested elsewhere on the forum, but in my opinion it should really be true. If you set it to `false`, you should always deduct the taker fee from the price, or GDAX will reject any order. I believe GDAX rejects it if it's (partially) covered in the book as a taker, in which case Gekko should automatically recalculate the price and try to place the order again. I haven't tested this last situation, but IMO that's what should happen.